refactor: add validation checks to WalletAppService methods

+++ src/BackEnd/go-services/WalletApi/internal/services/wallet_service.go
@@ -1,6 +1,7 @@
 package services
 
 import (
+	"fmt"
 	"google.golang.org/genproto/googleapis/type/decimal"
 	"time"
 	"wallet-api/internal/models"
@@ -40,27 +41,65 @@ type WalletAppService struct {
 func NewWalletService(repo repositories.WalletRepository) *WalletAppService {
 	return &WalletAppService{Repository: repo}
 }
-
 func (w *WalletAppService) Deposit(userID string, amount float64, idempotencyKey string) error {
+	if amount <= 0 {
+		return fmt.Errorf("deposit amount must be greater than zero")
+	}
+	if userID == "" {
+		return fmt.Errorf("user ID cannot be empty")
+	}
+	if idempotencyKey == "" {
+		return fmt.Errorf("idempotency key cannot be empty")
+	}
 	return w.Repository.Deposit(userID, amount, idempotencyKey)
 }
 
 func (w *WalletAppService) Withdraw(userID string, amount float64, idempotencyKey string) error {
+	if amount <= 0 {
+		return fmt.Errorf("withdraw amount must be greater than zero")
+	}
+	if userID == "" {
+		return fmt.Errorf("user ID cannot be empty")
+	}
+	if idempotencyKey == "" {
+		return fmt.Errorf("idempotency key cannot be empty")
+	}
+
+	currentAmount, err := w.Repository.GetAmount(userID)
+	if err != nil {
+		return fmt.Errorf("error retrieving current amount: %w", err)
+	}
+	if currentAmount < amount {
+		return fmt.Errorf("insufficient funds")
+	}
+
 	return w.Repository.Withdraw(userID, amount, idempotencyKey)
 }
 
 func (w *WalletAppService) GetTransactions(userID string) ([]models.Transaction, error) {
+	if userID == "" {
+		return nil, fmt.Errorf("user ID cannot be empty")
+	}
 	return w.Repository.GetTransactions(userID)
 }
 
 func (w *WalletAppService) GetAmount(userID string) (float64, error) {
+	if userID == "" {
+		return 0, fmt.Errorf("user ID cannot be empty")
+	}
 	return w.Repository.GetAmount(userID)
 }
 
 func (w *WalletAppService) RemoveExpire(createdAt time.Time) error {
+	if createdAt.IsZero() {
+		return fmt.Errorf("createdAt cannot be zero")
+	}
 	return w.Repository.RemoveExpire(createdAt)
 }
 
 func (w *WalletAppService) GetTransactionsReport(yearNumber int) (map[int]decimal.Decimal, error) {
+	if yearNumber <= 0 {
+		return nil, fmt.Errorf("year number must be greater than zero")
+	}
 	return w.Repository.GetTransactionsReport(yearNumber)
 }
